requirements.txt
Flask
Flask-Cors
mysql-connector-python
python-dotenv
flask_jwt_extended
bcrypt
requests
